[x/tour] content: add comment about << in range example
The 'Range Continued' code example uses a left shift operator.
Some people are unfamiliar with shift, and being puzzled about
what it means can distract from the point--which is range.

Per a suggestion from gri, this adds a small comment to show that
it's equivalent to 2**n.

Fixes golang/go#14241

Change-Id: If0bcfad85390db96f9f9c4295dc18f34410cb494
Reviewed-on: https://go-review.googlesource.com/19291
Reviewed-by: Robert Griesemer <gri@golang.org>
X-Tour-Commit: 840f2c28917e5e5f274195ba2d7d8434aee4399a