[x/tour] content: show actual value of numeric constants
The code example for the Numeric Constants slide uses the
right and left shift operators, which some people are
unfamiliar with.

Showing the values of the constants in a comment lets
people grasp the point of the slide without being distracted
by not understanding the shift operators.

Fixes golang/go#14151

Change-Id: I390655c0f9392f2e5b00d813fa21b8a226cf2f39
Reviewed-on: https://go-review.googlesource.com/19325
Reviewed-by: Andrew Gerrand <adg@golang.org>
X-Tour-Commit: cf672f18b3ea5584028440a6624c7c6c94c6d584