[x/tour] content: clarify that tree.New returns sorted trees
The description of the 'Equivalent Binary Trees' exercise says:

  The function tree.New constructs a randomly-structured binary tree

In the context of CS, 'binary trees' (in the 'data structure' sense)
are sorted by definition, but sometimes the term is used, expecially
in mathematics, to indicate the more general notion of a tree with at
most two children.

Since it has been reported that tour users are sometimes confused by
this (see the linked issue), clarify that the tree.New function
returns random, but *always sorted* trees.

Fixes golang/tour#270

Change-Id: Ibfca0e2a8e947a839545433d1da0f4621ea65340
Reviewed-on: https://go-review.googlesource.com/56171
Reviewed-by: Andrew Bonventre <andybons@golang.org>
Reviewed-by: Rob Phoenix <rob@robphoenix.com>
X-Tour-Commit: 279fc850d72ee755bf08339069eb9ebdf23a0620