[x/blog] blog: version of ISMM keynote
This is a blog containing the slides and
a transcript of the ISMM keynote that discusses
the work we have done to date on the Go GC.

It is meant to read like it was a talk and
attempts to maintain some of the informality of
the actual talk without sacrificing clarity.
It's not supposed to have the feel of a formal
academic paper.

Change-Id: Ied4ba8973cbd6a0808fe72e809ff8cbc565672a9
Reviewed-on: https://go-review.googlesource.com/122797
Reviewed-by: Cassandra Salisbury <cls@golang.org>
X-Blog-Commit: 8d4da3ad867c5dafb117ad39d8f9a13a1cc2d8fd