cmd/golangorg/godoc: delete packages
The godoc package and the packages inside (an exact copy of
x/tools/godoc) are being removed from x/website/cmd/golangorg
because the package should not be in the folder used to build
the binary when it's being used as a library. Also, all the work
to turn documentation-rendering into a library separate from
x/website and the godoc binary will be done in x/tools/godoc.
Over time files and packages that are solely used by the website
will be moved back into this repo. All corresponding path name
changes were made to all files importing the godoc/... packages.

Updates golang/go#29206

Change-Id: I13740b66b967fc413eed7e90bff54f859cf36150
Reviewed-on: https://go-review.googlesource.com/c/159737
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>