cmd/golangorg, internal/env: make deployment work for x/website
This change is part of an effort to deploy x/website to its own
subdomain on GCP. Everything needs to be working properly on the
subdomain before x/website can become the canonical website
(golang.org). There is currently a functional deployment of
x/website to the 'new-website' service under golang-org on GCP.

With these specific changes, x/website gets deployed to the
service 'new-website'. By changing both the service field in
app.prod.yaml and the GCP_SERVICE variable in Makefile to
'default', these files could be used to deploy x/website to
the 'default' service as well.

Updates golang/go#29206

Change-Id: I47340333b9e09672e8588c330d629857d1d0bf0d
Reviewed-on: https://go-review.googlesource.com/c/161198
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>