cmd/golangorg: link to a more narrow list of old versions
Update the GCP console link to have an Environment=Flexible filter,
so the list of old website versions doesn't include extraneous entries.

The versionsquery query parameter is not easily readable, but it's what
GCP uses. It encodes the following information:

	[{"k":"env","t":10,"v":"\"FLEXIBLE\"","s":true}]

Remove the versionssize=50 query parameter, since it's not critical,
to make the URL shorter and easier to copy.

Change-Id: I09e7c25191668af10aef0783b2760af3b7873654
Reviewed-on: https://go-review.googlesource.com/c/163177
Reviewed-by: Channing Kimble-Brown <channing@golang.org>