internal/env: use golangorgenv package from x/tools
This change de-duplicates the same configuration used between
the previous x/tools/godoc/env package and x/website's own
internal/env package. It does so by using the new golangorgenv
package.

This makes it possible to remove the old GODOC_PROD and
GODOC_ENFORCE_HOSTS env vars from app.prod.yaml.

The IsProd configuration value was very generic and yet being
used only in one place. Replace it with a more targeted one.
The new name does a better job of communicating what the
implications of setting or not setting it are.

Apply the same change to logic of googleCN as in CL 165459.

Update the required version of the x/tools module to a newer version
that includes CL 160837, which is required for this change.

Starting with this change, we're changing the x/website policy in that
it's now allowed to start using newer versions of subrepos from the
master branch, rather than only using the release-branches of subrepos.
This is in line with the goal of making x/website more self-contained
and less constrained by the Go release schedule.

Updates golang/go#29206

Change-Id: If8847e9276fb18e3ded56fabbf915cdd19176699
Reviewed-on: https://go-review.googlesource.com/c/website/+/165837
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>