content/static: revamp main header and clean up markup
This is the first of many changes to spiff up golang.org a bit
to include our updated branding and iterate on the design.

Updates the html to be more modern and rewrites the header CSS to
be in line with our style guide (golang.org/wiki/CSSStyleGuide).

Additionally, removes the play popup widget in the header nav
in favor of linking directly to the playground and formats
the CSS using prettier.

Updates golang/go#9936

Change-Id: I9baf727f9e703b560ef786277b5d9859baa6fbfc
Reviewed-on: https://go-review.googlesource.com/c/website/+/178137
Reviewed-by: Katie Hockman <katie@golang.org>