content/static, internal/dl: update header on downloads page
The downloads page does not share the root template as other pages.
Update it to use the new header markup.

Also fixes an issue where static.go did not contain the Go logo and
updates the downloads page to be consistent with others when visiting
from China.

Updates golang/go#9936
Fixes golang/go#32151

Change-Id: Ic6d703ccb8fe5cf93d74256bac41ecad69a78e72
Reviewed-on: https://go-review.googlesource.com/c/website/+/178138
Reviewed-by: Katie Hockman <katie@golang.org>