content/static: remove overflow: auto for header elements
A header element should not be forced to scroll when its height
is smaller than its children. Let it overflow instead.

overflow: auto is used to ensure that child floating elements
are contained within their parent. For the case where this is
used (Go version that a function was added), the version number
won’t be any larger than the other text in the header, that
declaration isn’t needed.

Fixes golang/go#32739

Change-Id: I99aea447b19c95840c04af7d7e13280af1ac05de
Reviewed-on: https://go-review.googlesource.com/c/website/+/183597
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>