content/static: show the "Play" link in the navigation header on the landing page
The "Play" link is visible when the page title is set, but it's not set on the landing page.
By this change, the "Play" link is visible no matter whether the page title is set.

Fixes golang/go#33043

Change-Id: I6a35e3d5c3d54ba5bd3d23967196cb7edaeefa22
Reviewed-on: https://go-review.googlesource.com/c/website/+/187897
Run-TryBot: Andrew Bonventre <andybons@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>