cmd/golangorg: remove autocert support
It was unused and incomplete.

It was potentially needed a while ago, back when the golang.org
website was running on a different Google Cloud Platform service.

It's not needed now, and can be re-added in the future
if it becomes needed again.

Change-Id: Id52945adf18b1e34f411f9a2eef6bd21889c6c1a
Reviewed-on: https://go-review.googlesource.com/c/website/+/197957
Run-TryBot: Dmitri Shuralyov <dmitshur@golang.org>
TryBot-Result: Gobot Gobot <gobot@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>