cmd/golangorg: remove -url flag
The -url flag is unused. It had relatively niche
use-cases and was mostly a relic from cmd/godoc.

It's also completely non-functional due to a few bugs and
the fact CL 162907 hasn't been backported to x/website.

Instead of fixing it and adding tests, remove it.
If it becomes needed in the future, it can be re-added.

Update golang/go#29206

Change-Id: I4a460f460a757c78c3e26d7c97556c623c8dbed4
Reviewed-on: https://go-review.googlesource.com/c/website/+/197722
Reviewed-by: Brad Fitzpatrick <bradfitz@golang.org>
Reviewed-by: Andrew Bonventre <andybons@golang.org>