cmd/golangorg: update GCP console UI link
The URL format has changed slightly, update it so that the Flexible
environment filter continues to be active.

Also, avoid printing directions at the end of the publish step in the
Makefile. Doing so creates overlap with the steps already listed in
the README and can be confusing. Instead, point back to the README.

Change-Id: If1d06735dff5744b81ad127e8621a35fa5bf167b
Reviewed-on: https://go-review.googlesource.com/c/website/+/234679
Reviewed-by: Chris Broadfoot <cbro@golang.org>