internal/dl: add includes values stable, unstable, archive
The include query parameter allows for one value, all, which includes all
stable, archived, and unstable releases in the response. The dl endpoint
supports getting a list of stable releases and a list of all stable, archived,
and unstable releases, but not getting other subsets of the releases. This
makes it non-trivial for users of the JSON endpoint to get a list of stable and
unstable releases together which would make it easier for folks auto-installing
Go versions to auto-install beta and release candidates alongside stable
release.

This change makes it possible to pass other values to the include query
parameter and to choose what types of releases are returned. It supports the
values stable, unstable, and archive, alongside the existing all value.