[x/blog] content: fix broken links and use HTTPS on godoc.article
The article contains links to code for strings.HasPrefix that reference
the wrong line number and URL to search for "Deprecated:" identifier
which also not working anymore.

This commit fix the line number link for strings.HasPrefix and
remove the line that describe examples for searching "Deprecated:"
identifier.

While at it, use HTTPS for link to Docstring and update the link
on Javadoc to latest location (according to the page redirect URL).

Signed-off-by: shulhan <m.shulhan@gmail.com>
Change-Id: Ife267af49c0816df2b3934d075c7a9a09eaac83e
Reviewed-on: https://go-review.googlesource.com/c/blog/+/302729
Reviewed-by: Ian Lance Taylor <iant@golang.org>
X-Blog-Commit: 181b0cc4321bcd631c6c87654dd4840acfdcd568