highlight all changed parts of the tutorial code
This change is a fix for what appears to be a typo.  In the modules tutorial, in [section "Return greetings for multiple people"](https://golang.org/doc/tutorial/greetings-multiple-people), the step 2 code block does not indicate all the changed sections.

Everything I highlighted needs to changed in order for the code to work as intended.  I found the highlights to be very helpful in following and transcribing the tutorial, so I noticed when they were incorrect.

Change-Id: I0e5ccdef4053b2f6a7a5688e5be9c3f6307c5486
GitHub-Last-Rev: 84a5c6e3ba06aa69d83af041b8f9edef2cea35f4
GitHub-Pull-Request: golang/website#49
Reviewed-on: https://go-review.googlesource.com/c/website/+/311329
Reviewed-by: Steve Traut <straut@google.com>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>