[x/go.dev] all: move long one-page layouts into the page content
Hugo Markdown files cannot contain templates, so all the template
work had to be done in the layouts, partials, or shortcodes trees.

But the new setup treats the .md files as templates generating
Markdown (which then generates HTML), so we can move the
complex templates into the pages where they are used, instead of
having an unnecessary redirect.

Change-Id: I37ba42aaa700ec67d28cf2adf1bfa59b72bc2b2a
X-GoDev-Commit: 7ee84490c012de79fda2ed0ed1929dc83237453b