[x/tour] content: remove link to old video showing GOPATH workflow
The final 'Where to Go from here...' page of the tour suggests, as
next step, to look at a YT screencast titled 'Writing, building,
installing, and testing Go code', or to read the 'How To Write Go
Code' document. The screencast is from 2012, so of course it shows the
old GOPATH workflow.

This change removes the link to the video, and leaves the suggestion
to read 'How To Write Go Code', to make sure beginners won't start
using the legacy GOPATH-based workflow.

Fixes golang/tour#1089

Change-Id: Ib826298127d8465055c85a2537732e61ff21cdb0
Reviewed-on: https://go-review.googlesource.com/c/tour/+/319429
Trust: Alberto Donizetti <alb.donizetti@gmail.com>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>
X-Tour-Commit: a278aee398d54cb9f07b7e8961ffcd5d88abfda5