[x/go.dev] cmd/internal/tmplfunc: copy of rsc.io/tmplfunc
Package tmplfunc makes it possible to invoke other templates as if they
were functions, which greatly streamlines the use of sub-templates.

Making a copy avoids a dependency on an unreviewed personal repo.
Perhaps at some point in the future the main template package will
provide something like this functionality, and then we can remove this.

Change-Id: I67fd7ea159a47e8620793f0aca85e77432a5c45c
X-GoDev-Commit: 209d65474a5a6b6c25c7a7ab8fa9537702c69070