[x/go.dev] all: replace aliases with redirects
Putting an "aliases" list into a page means that you cannot serve
any page at all until you have loaded every single page in the site,
because maybe one will list the URL you want to serve as an alias.

Turning it around, we can create pages for the URLs we want to
serve and simply mark them as redirects to the real pages.
Now you can serve any page by opening the file and looking at
what is there. There's no need to reload every single page in the
site just to answer a single request.

We made (and corrected) this mistake in the main site as well.
It is an obvious thing to do.

Replace aliases with redirects.

Change-Id: Ib31bf48e2a6fa6f384011381dcae6bbfcaf11f95
X-GoDev-Commit: 8dcbcab6097450c570ca42101c2ea3218a8f1c62