[x/go.dev] cmd/internal: add forked copies of Go 1.17+ text/template and html/template
The most pressing need is a fixed eq that will ship in Go 1.17,
which allows comparing an interface nil against a string, as in
(eq .notavailable "hello"). It also has pending Go 1.18 features
break/continue and short-circuit and/or, which are likely to
be useful. It has the io/fs-based APIs from Go 1.16 removed,
since App Engine does not yet support io/fs.

Since we already had a copy of text/template/parse as
tmplfunc/internal/parse, changed tmplfunc to use the
copy being imported here.

Change-Id: I4619db1846c3002de38ae26f277b12e7041883b9
X-GoDev-Commit: 6f3361e314ab1d286011d8fc26e8ea4482415ec7