blog: delete per-repo files, update README.md, go.mod
It would be nice to delete go.mod and let the directory
be part of the x/website module, but right now x/website
requires Go 1.16 and the blog runs on App Engine which
doesn't yet support Go 1.16. So the blog is actually requiring
an older, pre-Go 1.16 version of x/website.

Once App Engine updates to Go 1.16 we can take that next
step of making blog not a submodule.

Or perhaps before that we will get to the point of serving the
blog from the cmd/golangorg server.

One way or another, this will get even simpler.

Change-Id: I04f9b4d4433009c8b204acba9043b2d8a6261189
Reviewed-on: https://go-review.googlesource.com/c/website/+/322791
Trust: Russ Cox <rsc@golang.org>
Reviewed-by: Dmitri Shuralyov <dmitshur@golang.org>