internal/web: stop double-escaping of code in non-playground examples
4c9e549253faa360e399b19be93adb3ad820ca4e introduced an issue causing non-playground
code examples to be double-escaped, causing HTML escape sequences and elements
to appear in the code, such as the examples in crypto/rsa, as reported in golang/go#46839.
This change treats the code as `template.HTML` rather than `string` after running
through `(*Page).Node` (which formats the code as HTML) and doing some basic transformations,
preventing this double-escaping from occurring.

Fixes golang/go#46839