go.dev/_content/css: move header bar to edges of page, like footer
Right now the header is constrained to the middle 75.75em of the page.
That lines up nicely when content is also limited to the middle 75.75em,
but some pages will want to allow a wider viewport, and then it looks weird
to have the content starting to the left of the Go logo and continuing past
the right of the tabs.

Let the header expand to the left and right sides of the page
(up to a margin on wider devices), just like the footer.

Change-Id: Ic667ec8d23dfb06ab6c532caad6b507dd5ce6fe5
Reviewed-on: https://go-review.googlesource.com/c/website/+/364814
Trust: Russ Cox <rsc@golang.org>
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>