_content: adjust footer
Change footer About heading to link to /project.
Linking to /about was always mainly about pkg.go.dev,
and now it redirects to pkg.go.dev/about, so that's clearly wrong.
But we need a link to somewhere, and we also need a link to /project
(which used to a be a top-level heading on golang.org),
so it's a good match.

Add Standard Library to Packages section.

Add Help to Get Started section.
(Used to be a top-level heading on golang.org.)

For golang/go#40785 (but probably more to do there).
Fixes golang/go#49737.
Fixes golang/go#49768.

Change-Id: I4a0837aaaa89b3bd79bda7a55bcff3a5ef9e206b
Reviewed-on: https://go-review.googlesource.com/c/website/+/366577
Trust: Russ Cox <rsc@golang.org>
Run-TryBot: Russ Cox <rsc@golang.org>
TryBot-Result: Go Bot <gobot@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>