Use var name consistently
The TestHelloName test declares and initializes the name variable as the
subject of the test. In the example, the variable is used only for the
definition of the regular expression. I suggest to use the variable also as
the argument of the method to be tested and for the error message to be output.