internal/history: document future Go 1.18.1 and 1.17.9 releases
These releases have been pre-announced to include a PRIVATE security fix:
https://groups.google.com/g/golang-announce/c/vtbMjE04kPk/m/xE-FGxCXCAAJ.

The fix summary is intentionally left blank until the day of the release,
other than to mention that the issue affects the standard library as the
Go security policy permits disclosing.

In the future, the go.dev website can be improved to display this
information automatically so that it doesn't involve a manual CL.
The motivation for this change was to investigate that the recent
FixSummary type and template are flexible enough to support that.

Change-Id: I340e3b44440cecdb9b25c426f97cd8ff31c60acd
Reviewed-on: https://go-review.googlesource.com/c/website/+/398214
Trust: Dmitri Shuralyov <dmitshur@google.com>
Run-TryBot: Dmitri Shuralyov <dmitshur@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Reviewed-by: Julie Qiu <julieqiu@google.com>