_content/security: add policy.html
We want to use go.dev/security as a landing page for all things Go
security.

In order to do so, go.dev/security now redirects to
go.dev/security/policy.

Once the landing page is created and all references to go.dev/security
are changed, go.dev/security will be its own page, with a link to the
security policy.

Change-Id: I281c38d3322abca35c87d9655debe126cc84cd64
Reviewed-on: https://go-review.googlesource.com/c/website/+/398255
Reviewed-by: Roland Shoemaker <roland@golang.org>
Reviewed-by: Zvonimir Pavlinovic <zpavlinovic@google.com>