_content: fix colors for links in BackgroundQuotes
The colors for the link and the background color for the background
quotes are too similar, so the quotes do not appear. This change adds
style to the links in the background quotes so that they inherit the
colors.

Updates golang/go#52312

Change-Id: Ib7fa0b23ff194333f50c1346747ea06d25e09254
Reviewed-on: https://go-review.googlesource.com/c/website/+/400894
Run-TryBot: Suzy Mueller <suzmue@golang.org>
Reviewed-by: Nooras Saba‎ <saba@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>