internal/screentest: check http status before page tasks
When a page request returned a 404 unexpectedly
it would cause a testcase that waited for a css
selector to timeout instead of reporting the
test failure for mismatched status. This change
adds the status check as a task instead of a post
test check to report the failure after navigating
to a page.

Change-Id: I31c1b5dab6a412558c501d99873bff91d8dbc105
Reviewed-on: https://go-review.googlesource.com/c/website/+/409615
Reviewed-by: Jonathan Amsterdam <jba@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Auto-Submit: Jamal Carvalho <jamal@golang.org>