_content/doc/gc-guide: deemphasize STW collection in GC model
Based on feedback from aktau@google.com, it's more confusing to state it
as part of the model and then take it back later. At this point, it's
not really all that important to anything but the visualization, so this
CL moves the STW assumption to just the description of the
visualizations.

Change-Id: I345ae215e3ff06ad791044e00425472552df8022
Reviewed-on: https://go-review.googlesource.com/c/website/+/421054
Reviewed-by: Michael Pratt <mpratt@google.com>
Reviewed-by: Nicolas Hillegeer <aktau@google.com>