tour: no deterministic note as 1.20 does rand seed
With go 1.20 "math/rand package now automatically seeds the global random
number generator" (https://go.dev/doc/go1.20#math/rand) and we do not need
the appengine specific note that the Go Playground environment for the tour
programs is deterministic.

As the note is specific to the appengine version (i.e. it is already not shown
in the Go offline tour) it should be fine to just remove the text.

Fixes golang/tour#1448