tour: Reader has method not func
Interfaces do not define func as in the current text of "Readers"
("/tour/methods/21"). The interface specifies a list of methods, just Read() in
this case. It is confusing to say "interface has a Read method" and then show a
code line defining what would be a method on a type that impliments the
interface.

Changed the article text to follow the pattern of "Stringers"
("/tour/methods/17") which shows the interface definition in the code block. In
following this pattern I moved the code just below that text about the package
io and its io.Reader interface.

part 1/2