tour: mention generic functions in generics/1
The first slide of generics (about generic functions) never uses the
phrase "generic function". Adding to the confusion, the next slide
(generics/2) then says "In addition to generic functions..." as if the
reader should know about them.

Added the name "generic function" to generics/1 to explain that adding
type parameters to a func make a generic function.