tour: let reader know Walk should send in order
In "Exercise: Equivalent Binary Trees" neither the slide text nor the
code state Walk() should send the values in order. Doing so in order
makes using the output much more useful in Same(), but might not be easy
to see for readers that have not had a data structures class. In order
traversal is implied by reference with the binary tree being "always
sorted" and the output defined as "should be the numbers 1, 2, 3, ...,
10."

Update the slide text to say what Walk() does. Change Walk() comments in
code/solution to match.

Fixes golang/tour#158