tour: struct literals in same def/print order
The current code on "tour/moretypes/5" ("Struct Literals") seems to be
designed to confuse a reader into thinking that contents of 'v1' and 'p'
are importantly the same.  The contents of both variable's structs have
the same int values and then when printed they are (what seems like
purposefully, but with no explanation) printed next to each other.

Set the values in 'p' to be different and print the values in the same
order they are defined in.