tour: playground is always on a Tuesday
In "tour/flowcontrol/10" ("Switch evaluation order") the interaction
between the code using "Now()" and the Go Playground starting on
2009-11-10 is less than obvious.  The data the reader needs is that it
starts on a Tuesday.

Add the day of week to the online Go Playground note for readers to be
able to understand the output.

Fixes golang/tour#1407
Fixes golang/tour#1339
Fixes golang/tour#1249
Fixes golang/tour#1068