tour: clear up source/destination slice confusion
In "tour/moretypes/10" ("Slice defaults") the phrase "the length of the
slice for the high bound" has generated more than a dozen issues, mostly
wanting to replace "slice" with "array". The confusion is that when
slicing, the word "slice" is not well defined we will have an output
slice (result of the slice expression) and may well have a input slice
(e.g. in the example code in "slice-bounds.go"). The spec uses the
phrase "length of the sliced operand" for the high bound.

Change wording of slice default upper bound to phrase used in "ref/spec"
and reword the slide text to point out what the "sliced operands" are in
the slide and program examples.

Fixes golang/tour#1386
Fixes golang/tour#879
Fixes golang/tour#855
Fixes golang/tour#821
Fixes golang/tour#679
Fixes golang/tour#448
Fixes golang/tour#182

See golang/tour#182 for context and links to many of the issues.