tour: update links to go.dev
There are a few links in the tour (and tour source) that point at
golang.org rather than go.dev in places that are redirected to go.dev.
Seems reasonable to update these links as the public face for all of
these is go.dev now.

Some of these are in the headers of the *.article files, so I don't
think they show up anywhere on the web site, changed for consistency.
One was a blog link, just used what the web servers sent as the
redirect. Last pair were updating the links about the Go Playground, as
they are both behind "#appengine:" it is safe to assume access on go.dev
and use origin-relative links (e.g. /play/)

Update links to go.dev versions in tour article files.

Fixes golang/tour#1394