tour: document modern way to start offline tour
In 2018 tour/README.md got a update to look like the then existing
blog/README.md and the section got accidently placed under the patching
header (wrong location) and lost the header it had in blog/README.md. At
the time it was also still a python app.yaml command which, later
updated to be a "go run .", which no longer matches with the install
instructions for the tour which generate a "tour" binary file to run.

Add "Running Locally" header in (matching old blog/README.md style)
after the "Download/Install" and move local run instructions to that new
section while changing command to "tour" that is installed.