_content: add brand guidelines page: https://go.dev/brand
In 2018 Google filed for a trademark on the new Go Logo
(https://go.dev/blog/go-brand) and the word Go in the context of
programming languages (more specifically "Computer programs and
downloadable computer programs, namely compilers and reference
libraries, that implement a statically typed, compiled computer
programming language for use in developing, building and managing
other software.")

Since then we have not provided clear guidance about the permitted
uses of the logo and the word mark Go. This commit adds that guidance,
written by Google's trademark counsel and based on the Open Usage
Trademark Guidelines 0.1.0 [1], adapted for the Go project. We also
adapted some high-level text from the Python Software Foundation's
excellent Trademark Usage Policy [2]. The guidance in this commit will
be served at https://go.dev/brand and linked at the bottom of the menu
under "Brand Guidance" where there is currently a link to the blog
post.

As it says in the new web page, in general, we want the word mark “Go”
and the Go Logo to be used freely to refer to the Go programming
language. It is explicitly not a goal to restrict the many excellent
uses of the word Go in projects today. Instead the goal is to
authorize those excellent uses properly, to avoid any confusion or
doubt about whether they are proper.

As it also says in the new web page, anyone with questions can email
them to trademark@golang.org, which forwards to me and a few other
people at Google, and we will do our best to answer them.

[1] https://openusage.org/trademark-guidelines/
[2] https://www.python.org/psf/trademarks/

Fixes golang/go#58412.

Change-Id: I82aca152a4f33140b6f05eacd256de0d0ee686e9
Reviewed-on: https://go-review.googlesource.com/c/website/+/506756
Run-TryBot: Russ Cox <rsc@golang.org>
Reviewed-by: Bryan Mills <bcmills@google.com>
TryBot-Result: Gopher Robot <gobot@golang.org>
Auto-Submit: Russ Cox <rsc@golang.org>