_content/doc/contribute: make it clear a Gerrit account is required even when sending a GitHub PR
It's possible to read the current "Sending a change via GitHub"
section of the Contribution Guide as saying a Gerrit account
is optional when sending a GitHub PR.

Some of the friction listed by Bryan in discussion golang/go#61182
is likely the result of some GitHub users never registering for
Gerrit after they send their PR, such as:

  "A significant fraction of GitHub PRs are abandoned when
  the author does not respond to Gerrit comments"

This CL therefore attempts to make it clear a Gerrit account
is a requirement even when sending a PR via GitHub, which
might reduce a modest amount of friction.

See golang/go#61323 for additional background.

While we are here, we also suggest the PR author familiarize themselves
with Gerrit, which increases the likelihood of success for their
own PR, but also might help them eventually graduate to a pure Gerrit
workflow for future contributions or possibly open their eyes
to the benefits of doing basic gardening.

For example, starring and then following interesting CLs is a
great way to learn for a new contributor (but not as well known
a feature as it could be).

In addition, we add a link to the FAQs for the GitHub PR workflow,
which covers things like "why do I need a Google account for Gerrit?".

We also define the term "CL" and then purposefully use it. The
contribution guide currently does not use that term, but
some evidence suggests someone going through the GitHub PR workflow
can struggle a bit with the difference between the PR vs. the CL,
or generally be confused and ask questions like "what is a CL?".

(The Contribution Guide previously used the term "CL", but without
defining it. golang/go#18989 suggested defining the term,
though the resolution of that issue in CL 36564 instead
removed "CL" entirely from the contribution guide).

Fixes golang/go#61323

Change-Id: I7d16a8cfd357ec4dd441a6ca43fc2a0caf55cdfe
GitHub-Last-Rev: ce92b735e869f627e4bfa084b0ca721f93ef34ae
GitHub-Pull-Request: golang/website#232
Reviewed-on: https://go-review.googlesource.com/c/website/+/509019
Run-TryBot: t hepudds <thepudds1460@gmail.com>
Reviewed-by: Alan Donovan <adonovan@google.com>
Reviewed-by: Cherry Mui <cherryyz@google.com>
Reviewed-by: Ian Lance Taylor <iant@golang.org>
TryBot-Result: Gopher Robot <gobot@golang.org>