_content: fix error preventing some features to load in dl page
There was an error when setting up the callback for page load in the
/dl page, preventing some features to load properly. The error was due
to some methods being invoked for the playground textarea field, that
for this specific page is not in the dom.

Updated the logic for page load to wait for dom content loaded or check
document.readyState since the page might be loaded already when we try
to set up the callback.

Fixes: #61587