_content: fix error preventing some features to load in dl page
There was an error when setting up the callback for page load in the /dl page, preventing some features to load properly. The error was due to some methods being invoked for the playground textarea field, that for this specific page is not in the dom.

Updated the logic for page load to wait for dom content loaded or check document.readyState since the page might be loaded already when we try to set up the callback.

Fixes: golang/go#61587

Change-Id: I82508a2661167af19b1b4375c66b46ae50c08617
GitHub-Last-Rev: 8415de173b1edab071a65742e7193877d704ed89
GitHub-Pull-Request: golang/website#234
Reviewed-on: https://go-review.googlesource.com/c/website/+/513315
TryBot-Result: Gopher Robot <gobot@golang.org>
Run-TryBot: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Jamal Carvalho <jamal@golang.org>
Reviewed-by: Hyang-Ah Hana Kim <hyangah@gmail.com>