_content/doc/comment: mention "reports whether" idiom
Mention the idiom of using "reports whether" to describe functions
that return a boolean.

While here, replace several uses of "func" with "function". In the Go
spec, "func" is a keyword used in function declarations and function
literals, but the declared object is always referred to as a
"function".

Also, change "arguments" to "parameters". Again, in the Go spec,
"arguments" is used to describe the operands that appear in a call
expression, whereas "parameters" are the objects declared in a
function signature.

Change-Id: I01dfafd05af1852a8d2056bdabbda1e2bcc8cbb7
Reviewed-on: https://go-review.googlesource.com/c/website/+/524956
Reviewed-by: Rob Pike <r@golang.org>
Reviewed-by: Ian Lance Taylor <iant@google.com>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>