_content/doc/codewalk: fix code iframe size
This was broken at some point in the transition to the new site.
The iframe sizer was failing to run because '#footer' was not resolving.
If we fix it, then we have the problem that the "footer" is enormous
and we don't want to leave room for it on the screen anymore,
so just zero out that part of the computation.

This makes the codewalks viewable again.

Change-Id: I4080063bdb8988eeaf786e0f579e6e0877d97cb5
Reviewed-on: https://go-review.googlesource.com/c/website/+/534859
Auto-Submit: Russ Cox <rsc@golang.org>
LUCI-TryBot-Result: Go LUCI <golang-scoped@luci-project-accounts.iam.gserviceaccount.com>
Reviewed-by: Eli Bendersky <eliben@google.com>